op-node: change default L1 cache size to 900 (#14045)
In #13772, the L1 cache size was made configurable and the cap of 1000
was removed. This implied that the old default value (before applying
the cap) of 2/3 the sequencing window was now active.

However, high-throughput chains don't need such a high cache size under
normal conditions, so memory usage unexpectedly jumped for most node
operators.

This PR adds back a lower default value of 900, which is a better
default for most nodes. Only low-throughput chains with infrequent safe
head increments can now use this flag to set a higher value.