custom gas token: remove e2e test (#14094)
This commit removes an e2e test that covers the custom gas token
feature. Keeping this around will just result in extra maintenance
overhead that is not necessary.

Useful alongside:
- https://github.com/ethereum-optimism/optimism/pull/13603
- https://github.com/ethereum-optimism/optimism/pull/13921

for maintaining a patch that contains custom gas token. It should be
easy to bring this back in the future if necessary