fix(kt-devnet): make docker images idempotent (#14410)
Provenance attestations contain, among other things, a timestamp for the
build, which immediately makes the build non-reproducible.
This in turn creates unnecessary churn for the devnet, as each image
appears different even when it has the exact same content.

Conversely, skipping provenance is not really an issue in a devnet
environment.