op-program: Consolidate cascading block replacements (#14622)
* op-program: Consolidate cascading block replacements

Blocks containing invalid messages affect the cross-safety of its
dependencies. This is because once the block is replaced with a deposits-only block,
any messages it contained would no longer exist. And thus,
executing messages that referenced the non-existent messages become
invalid as well.

To deal with this, consolidation runs until all cross-depedencies are
reconciled against deposits-only updates.

* consolidate oracle

* op-e2e: intra-block cascade action case

* add unit tests for consolidate oracle

* always store state diff for the deposits-only block build

* add output root to block build db

* updated transition state view in consolidate oracle

* lint