fix(devnet-sdk): allow in-place redeployment (#14412)
Currently, the last step of the deployment (storing the devnet
descriptor into the enclave) fails the 2nd time around.
As kurtosis API doesn't allow us to overwrite an artifact, we sidestep
the issue by storing successive generations of the descriptor.

Also adjust the environment fetcher to detect and load the last
generation.