op-e2e/actions: test batcher should not warn when safe head moves (#14912)
This implementation of a batcher does not have block submission tracking fully implemented. Therefore it is expected that the "last submitted block" won't move before the safe head on the sequencer moves.

This change just marks that as expected behaviour to avoid sending devs on a wild goose chase when debugging.